move AddTask to cxx, for ROOT6 compatibility
