+++ PWGJE/EMCALJetTasks/macros/AddTaskEmcalJetPerformance.C
@@ -4,137 +4,12 @@ AliAnalysisTaskEmcalJetPerformance* AddTaskEmcalJetPerformance(
   const char *nGenLev            = "mcparticles",
   const Double_t minTrPt         = 0.15,              // Minimum track pT in standard track container
   const Double_t minClPt         = 0.30,              // Minimum cluster E in standard cluster container
-  const char *suffix             = ""
-)
+  const char *suffix             = "")
 {
-  // Get the pointer to the existing analysis manager via the static access method.
-  //==============================================================================
-  AliAnalysisManager *mgr = AliAnalysisManager::GetAnalysisManager();
-  if (!mgr)
-  {
-    ::Error("AddTaskEmcalJetPerformance", "No analysis manager to connect to.");
-    return 0;
-  }
-
-  // Check the analysis type using the event handlers connected to the analysis manager.
-  //==============================================================================
-  AliVEventHandler* handler = mgr->GetInputEventHandler();
-  if (!handler)
-  {
-    ::Error("AddTaskEmcalJetPerformance", "This task requires an input event handler");
-    return 0;
-  }
-
-  enum EDataType_t {
-    kUnknown,
-    kESD,
-    kAOD
-  };
-
-  EDataType_t dataType = kUnknown;
-
-  if (handler->InheritsFrom("AliESDInputHandler")) {
-    dataType = kESD;
-  }
-  else if (handler->InheritsFrom("AliAODInputHandler")) {
-    dataType = kAOD;
-  }
-
-  //-------------------------------------------------------
-  // Init the task and do settings
-  //-------------------------------------------------------
-
-  TString trackName(ntracks);
-  TString clusName(nclusters);
-
-  if (trackName == "usedefault") {
-    if (dataType == kESD) {
-      trackName = "Tracks";
-    }
-    else if (dataType == kAOD) {
-      trackName = "tracks";
-    }
-    else {
-      trackName = "";
-    }
-  }
-
-  if (clusName == "usedefault") {
-    if (dataType == kESD) {
-      clusName = "CaloClusters";
-    }
-    else if (dataType == kAOD) {
-      clusName = "caloClusters";
-    }
-    else {
-      clusName = "";
-    }
-  }
-
-  TString name("AliAnalysisTaskEmcalJetPerformance");
-  if (!trackName.IsNull()) {
-    name += "_";
-    name += trackName;
-  }
-  if (!clusName.IsNull()) {
-    name += "_";
-    name += clusName;
-  }
-  if (strcmp(suffix,"") != 0) {
-    name += "_";
-    name += suffix;
-  }
-
-  /////////////////////////////////////////////////////////////
-  // Configure di-jet task
-  AliAnalysisTaskEmcalJetPerformance* task = new AliAnalysisTaskEmcalJetPerformance(name);
-  
-  /////////////////////////////////////////////////////////////
-  // Create track and cluster containers with the standard cuts
-  
-  AliParticleContainer* partCont = 0;
-  if (trackName == "mcparticles") {
-    AliMCParticleContainer* mcpartCont = new AliMCParticleContainer(trackName);
-    partCont = mcpartCont;
-  }
-  else if (trackName == "tracks" || trackName == "Tracks") {
-    AliTrackContainer* trackCont = new AliTrackContainer(trackName);
-    partCont = trackCont;
-  }
-  if (partCont) partCont->SetParticlePtCut(minTrPt);
-  if (partCont) task->AdoptParticleContainer(partCont);
-  
-  // Add the generator-level container, if specified
-  if (nGenLev && strcmp(nGenLev,"")!=0) {
-    AliMCParticleContainer* mcpartCont = task->AddMCParticleContainer(nGenLev);
-    mcpartCont->SelectPhysicalPrimaries(kTRUE);
-    mcpartCont->SetParticlePtCut(0);
-  }
-  
-  // Add the cluster container
-  AliClusterContainer* clusCont = 0;
-  if (!clusName.IsNull()) {
-    clusCont = new AliClusterContainer(clusName);
-    clusCont->SetClusECut(0.);
-    clusCont->SetClusPtCut(0.);
-  }
-  if (clusCont) task->AdoptClusterContainer(clusCont);
-
-  //-------------------------------------------------------
-  // Final settings, pass to manager and set the containers
-  //-------------------------------------------------------
-
-  mgr->AddTask(task);
-
-  // Create containers for input/output
-  AliAnalysisDataContainer *cinput1  = mgr->GetCommonInputContainer()  ;
-  TString contname(name);
-  contname += "_histos";
-  AliAnalysisDataContainer *coutput1 = mgr->CreateContainer(contname.Data(),
-      TList::Class(),AliAnalysisManager::kOutputContainer,
-      Form("%s", AliAnalysisManager::GetCommonFileName()));
-  mgr->ConnectInput  (task, 0,  cinput1 );
-  mgr->ConnectOutput (task, 1, coutput1 );
-
-  return task;
+  return AliAnalysisTaskEmcalJetPerformance::AddTaskEmcalJetPerformance(ntracks,
+                                                                        nclusters,
+                                                                        nGenLev,
+                                                                        minTrPt,
+                                                                        minClPt,
+                                                                        suffix);
 }
